hngx_stage1-main (1)/hngx_stage1-main/requirements.txt
attrs==23.1.0
click==8.1.7
flasgger==0.9.7.1
Flask==2.2.3
Flask-Cors==3.0.10
itsdangerous==2.1.2
Jinja2==3.1.2
jsonschema==4.19.0
jsonschema-specifications==2023.7.1
MarkupSafe==2.1.3
mistune==3.0.1
packaging==23.1
python-dotenv==1.0.0
PyYAML==6.0.1
referencing==0.30.2
rpds-py==0.10.2
six==1.16.0
Werkzeug==2.3.7
